mergerbug without goto

+++ physical_operator/exchange_merger.h
@@ -171,7 +171,7 @@ class ExchangeMerger : public PhysicalOperator {
   PerformanceInfo* perf_info_;
   bool is_registered_to_tracker_;
 
-  int confirm_sender_time;
+  double confirm_sender_time;
   int frequence;
 
  private:
